nits

+++ script/run-gpu.sh
@@ -1,7 +1,7 @@
 echo '** This is GPU server script **'
 sed -i 's/1/0/g' signal
-# model=("vggsoter" "vggennclave" "vggag" "mlcapsule" "vggalex")
-model=("vggalex")
+# model=("vggsoter" "vggennclave" "vggag" "mlcapsule" "alexsoter")
+model=("alexsoter")
 for ((i = 0 ; i < ${#model[@]} ; i++))
 do
     if [ "${model[$i]}" == "vggsoter" ];then
@@ -115,7 +115,7 @@ do
             fi
         done
     fi
-    if [ "${model[$i]}" == "vggalex" ];then
+    if [ "${model[$i]}" == "alexsoter" ];then
         var=$(head -n +1 signal)
         echo -n "[Before running] signal =  "
         echo $var
@@ -131,6 +131,35 @@ do
             then
                 echo "[After running] Signal reset to 1 by relay script"
                 ps aux | grep gpu_server | awk 'NR==1{print $2}'
+                proc=$(ps aux | grep gpu_server | awk 'NR==1{print $2}')
+                kill -9 $proc  
+                echo "[After running] VGG completed. Process killed. Exit"
+                sed -i 's/1/0/g' signal
+                var=$(head -n +1 signal)
+                if [ "$var" == "0" ];then
+                    echo "[After running] Signal reset to 0 by gpu script"  
+                fi 
+                break
+            else 
+                echo "[After running] Signal = 0 "  
+            fi
+        done
+    fi
+    if [ "${model[$i]}" == "alexag" ];then
+        var=$(head -n +1 signal)
+        echo -n "[Before running] signal =  "
+        echo $var
+        pushd /home/xian/atc22-artifact/SOTER/other-partition/aegisdnn/alex-partition/cpp/ag-graphene-an
+        bash runserver.sh
+        popd
+        ssh jianyu@10.22.1.16 "bash ~/atc22-artifact/SOTER/script/run-relay-forw.sh ${model[$i]}"
+        while :
+        do
+            sleep 1s
+            var=$(head -n +1 signal)
+            if [ "$var" == "1" ]
+            then
+                echo "[After running] Signal reset to 1 by relay script"
                 proc=$(ps aux | grep gpu_server | awk 'NR==1{print $2}')
                 kill -9 $proc  
                 echo "[After running] VGG completed. Process killed. Exit"
